# Optimize expression evaluation and eliminate allocations