--- a/src/NRules/NRules/ActionInvocation.cs
+++ b/src/NRules/NRules/ActionInvocation.cs
@@ -17,7 +17,7 @@ namespace NRules
             _action = action;
         }
 
-        public object[] Arguments => null;
+        public object[] Arguments => _action.GetArguments(_actionContext);
         public Expression Expression => _action.Expression;
         public ActionTrigger Trigger => _action.Trigger;