--- a/src/NRules/NRules/AgendaFilters/ActivationExpression.cs
+++ b/src/NRules/NRules/AgendaFilters/ActivationExpression.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Linq.Expressions;
-using NRules.Rete;
-using NRules.Utilities;
+using Tuple = NRules.Rete.Tuple;
 
 namespace NRules.AgendaFilters
 {
@@ -13,50 +12,34 @@ namespace NRules.AgendaFilters
     internal class ActivationExpression<TResult> : IActivationExpression<TResult>
     {
         private readonly LambdaExpression _expression;
-        private readonly FastDelegate<Func<object[], TResult>> _compiledExpression;
-        private readonly IndexMap _expressionFactMap;
+        private readonly Func<Tuple, TResult> _compiledExpression;
 
-        public ActivationExpression(LambdaExpression expression, FastDelegate<Func<object[], TResult>> compiledExpression, IndexMap expressionFactMap)
+        public ActivationExpression(LambdaExpression expression, Func<Tuple, TResult> compiledExpression)
         {
             _expression = expression;
             _compiledExpression = compiledExpression;
-            _expressionFactMap = expressionFactMap;
         }
 
         public TResult Invoke(AgendaContext context, Activation activation)
         {
-            var tuple = activation.Tuple;
-            var tupleFactMap = activation.FactMap;
-
-            var args = new object[_compiledExpression.ArrayArgumentCount];
-
-            int index = tuple.Count - 1;
-            var enumerator = tuple.GetEnumerator();
-            while (enumerator.MoveNext())
-            {
-                var mappedIndex = _expressionFactMap[tupleFactMap[index]];
-                IndexMap.SetElementAt(args, mappedIndex, enumerator.Current.Object);
-                index--;
-            }
-
             Exception exception = null;
             TResult result = default;
             try
             {
-                result = _compiledExpression.Delegate.Invoke(args);
+                result = _compiledExpression.Invoke(activation.Tuple);
                 return result;
             }
             catch (Exception e)
             {
                 exception = e;
                 bool isHandled = false;
-                context.EventAggregator.RaiseAgendaExpressionFailed(context.Session, e, _expression, args, activation, ref isHandled);
+                context.EventAggregator.RaiseAgendaExpressionFailed(context.Session, e, _expression, null, activation, ref isHandled);
                 throw new ExpressionEvaluationException(e, _expression, isHandled);
             }
             finally
             {
                 if (context.EventAggregator.TraceEnabled)
-                    context.EventAggregator.RaiseAgendaExpressionEvaluated(context.Session, exception, _expression, args, result, activation);
+                    context.EventAggregator.RaiseAgendaExpressionEvaluated(context.Session, exception, _expression, null, result, activation);
             }
         }
     }