--- a/src/NRules/NRules/AgendaFilters/ActivationExpression.cs
+++ b/src/NRules/NRules/AgendaFilters/ActivationExpression.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Linq.Expressions;
+using NRules.Utilities;
 using Tuple = NRules.Rete.Tuple;
 
 namespace NRules.AgendaFilters
@@ -13,11 +14,13 @@ namespace NRules.AgendaFilters
     {
         private readonly LambdaExpression _expression;
         private readonly Func<Tuple, TResult> _compiledExpression;
+        private readonly IArgumentMap _argumentMap;
 
-        public ActivationExpression(LambdaExpression expression, Func<Tuple, TResult> compiledExpression)
+        public ActivationExpression(LambdaExpression expression, Func<Tuple, TResult> compiledExpression, IArgumentMap argumentMap)
         {
             _expression = expression;
             _compiledExpression = compiledExpression;
+            _argumentMap = argumentMap;
         }
 
         public TResult Invoke(AgendaContext context, Activation activation)
@@ -33,13 +36,13 @@ namespace NRules.AgendaFilters
             {
                 exception = e;
                 bool isHandled = false;
-                context.EventAggregator.RaiseAgendaExpressionFailed(context.Session, e, _expression, null, activation, ref isHandled);
+                context.EventAggregator.RaiseAgendaExpressionFailed(context.Session, e, _expression, _argumentMap, activation, ref isHandled);
                 throw new ExpressionEvaluationException(e, _expression, isHandled);
             }
             finally
             {
                 if (context.EventAggregator.TraceEnabled)
-                    context.EventAggregator.RaiseAgendaExpressionEvaluated(context.Session, exception, _expression, null, result, activation);
+                    context.EventAggregator.RaiseAgendaExpressionEvaluated(context.Session, exception, _expression, _argumentMap, result, activation);
             }
         }
     }