--- a/src/NRules/NRules/Diagnostics/AgendaExpressionErrorEventArgs.cs
+++ b/src/NRules/NRules/Diagnostics/AgendaExpressionErrorEventArgs.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Linq.Expressions;
 using NRules.RuleModel;
+using NRules.Utilities;
 
 namespace NRules.Diagnostics
 {
@@ -17,6 +18,11 @@ namespace NRules.Diagnostics
         /// <param name="arguments">Expression arguments.</param>
         /// <param name="match">Rule match related to the event.</param>
         public AgendaExpressionErrorEventArgs(Expression expression, Exception exception, object[] arguments, IMatch match)
+            : base(expression, exception, arguments, null, match)
+        {
+        }
+        
+        internal AgendaExpressionErrorEventArgs(Expression expression, Exception exception, IArguments arguments, IMatch match)
             : base(expression, exception, arguments, null, match)
         {
         }