--- a/src/NRules/NRules/Diagnostics/LhsExpressionErrorEventArgs.cs
+++ b/src/NRules/NRules/Diagnostics/LhsExpressionErrorEventArgs.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq.Expressions;
 using NRules.RuleModel;
+using NRules.Utilities;
 
 namespace NRules.Diagnostics
 {
@@ -24,17 +25,8 @@ namespace NRules.Diagnostics
         {
         }
         
-        /// <summary>
-        /// Initializes a new instance of the <c>LhsExpressionErrorEventArgs</c> class.
-        /// </summary>
-        /// <param name="expression">Expression related to the event.</param>
-        /// <param name="exception">Exception related to the event.</param>
-        /// <param name="argument">Expression argument.</param>
-        /// <param name="tuple">Tuple related to the event.</param>
-        /// <param name="fact">Fact related to the event.</param>
-        /// <param name="rules">Rules that contain the expression that generated the event.</param>
-        public LhsExpressionErrorEventArgs(Expression expression, Exception exception, object argument, ITuple tuple, IFact fact, IEnumerable<IRuleDefinition> rules)
-            : base(expression, exception, argument, null, tuple, fact, rules)
+        internal LhsExpressionErrorEventArgs(Expression expression, Exception exception, IArguments arguments, ITuple tuple, IFact fact, IEnumerable<IRuleDefinition> rules)
+            : base(expression, exception, arguments, null, tuple, fact, rules)
         {
         }