… bench/NRules.Benchmark/NRules.Benchmark/Expressions/BenchmarkLhsExpression.cs
… bench/NRules.Benchmark/NRules.Benchmark/Expressions/BenchmarkRuleAction.cs
… src/NRules/NRules/Diagnostics/RhsExpressionErrorEventArgs.cs
… src/NRules/NRules/Diagnostics/RhsExpressionEventArgs.cs
… src/NRules/NRules/Diagnostics/SessionSnapshotVisitor.cs
… src/NRules/NRules/Rete/LhsExpression.cs
… src/NRules/NRules/Rete/ReteBuilder.cs
… src/NRules/NRules/Rete/ReteBuilderContext.cs
… src/NRules/NRules/Rete/ReteNodeVisitor.cs
… src/NRules/NRules/Rete/RuleNode.cs
… src/NRules/NRules/Rete/Terminal.cs
… src/NRules/NRules/Rete/TerminalNode.cs
… src/NRules/NRules/Rete/Tuple.cs
… src/NRules/NRules/RuleAction.cs
… src/NRules/NRules/RuleCompiler.cs
… src/NRules/NRules/RuleDependency.cs
… src/NRules/NRules/Utilities/ArgumentMap.cs
… src/NRules/NRules/Utilities/Arguments.cs
… src/NRules/NRules/Utilities/ExpressionCompiler.cs
… src/NRules/NRules/Utilities/ExpressionOptimizer.cs
… src/NRules/NRules/Utilities/FastDelegate.cs
… src/NRules/Tests/NRules.IntegrationTests/EvaluationEventTest.cs
… src/NRules/Tests/NRules.IntegrationTests/ExpressionParameterOrderTest.cs
… src/NRules/Tests/NRules.IntegrationTests/OneFactRuleWithDependencyTest.cs
… src/NRules/Tests/NRules.Tests/AgendaTest.cs
… src/NRules/Tests/NRules.Tests/Rete/TupleTest.cs
… src/NRules/Tests/NRules.Tests/SessionTest.cs