This is a change for #200 to eliminate memory allocations when evaluating expressions, as well as to optimize the performance of expression evaluation.
The gist of the problem is this - how to invoke a dynamically compiled lambda expression with a variable number of arguments. Dynamic invoke is too slow. The original solution rewrote a lambda expression from `Func<p0, p1, p2, ..., pn, result>` to `Func<p[n], result>` thereby making a single array argument expression. But this required allocating an array prior to every single expression evaluation, and populating the array with the values of parameters.
This change modifies the approach and rewrites the expression from `Func<p0, p1, p2, ..., pn, result>` to a variation of `Func<Tuple, Fact, result>` thereby weaving the unwrapping of the tuple directly into the expression code, setting expression arguments on the stack, eliminating memory allocations.
There are additional improvements that further sped up the performance of expression evaluation. E.g. changing Tuple enumerator from a class to a struct, eliminating one more allocation.
One challenge this improvement presented is around the expression evaluation diagnostics, where previously, expression arguments were available as an `object[]` on every evaluation or failure event, as well as on `IActionInvocation`. This required that the arguments were packaged as object arrays, which is no longer the case.
To preserve the functionality, instead of sending argument arrays on the diagnostics events, a lazy iterator is sent that allocates the array and unwraps the corresponding tuple only when the client requests the arguments. Therefore, the client only pays the cost of allocation and tuple unwrapping if they are using the arguments.