@snikolayev 2020-05-24
Before:
``` ini
BenchmarkDotNet=v0.12.1, OS=Windows 10.0.18363.836 (1909/November2018Update/19H2)
Intel Core i7-10710U CPU 1.10GHz, 1 CPU, 12 logical and 6 physical cores
  [Host]     : .NET Framework 4.8 (4.8.4180.0), X64 RyuJIT
  DefaultJob : .NET Framework 4.8 (4.8.4180.0), X64 RyuJIT
```
|                      Method |     Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------------------- |---------:|---------:|---------:|-------:|------:|------:|----------:|
| EvaluateTupleFactExpression | 77.60 ns | 1.444 ns | 1.280 ns | 0.0191 |     - |     - |     120 B |
|     EvaluateTupleExpression | 80.37 ns | 1.238 ns | 1.097 ns | 0.0191 |     - |     - |     120 B |
|      EvaluateFactExpression | 15.71 ns | 0.334 ns | 0.411 ns | 0.0038 |     - |     - |      24 B |
| EvaluateExpression | 85.40 ns | 1.683 ns | 1.938 ns | 0.0153 |     - |     - |      96 B |

After:
``` ini
BenchmarkDotNet=v0.12.1, OS=Windows 10.0.18363.836 (1909/November2018Update/19H2)
Intel Core i7-10710U CPU 1.10GHz, 1 CPU, 12 logical and 6 physical cores
  [Host]     : .NET Framework 4.8 (4.8.4180.0), X64 RyuJIT
  DefaultJob : .NET Framework 4.8 (4.8.4180.0), X64 RyuJIT
```
|                      Method |     Mean |    Error |   StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------------------- |---------:|---------:|---------:|------:|------:|------:|----------:|
| EvaluateTupleFactExpression | 23.37 ns | 0.211 ns | 0.176 ns |     - |     - |     - |         - |
|     EvaluateTupleExpression | 29.10 ns | 0.052 ns | 0.049 ns |     - |     - |     - |         - |
|      EvaluateFactExpression | 11.38 ns | 0.175 ns | 0.136 ns |     - |     - |     - |         - |
| EvaluateExpression | 39.75 ns | 0.814 ns | 0.762 ns |     - |     - |     - |         - |